requirements.txt
pandas-datareader
python-telegram-bot